Fix Track detail fragment delete and ActionBar
When showing the details for a track we remove the up navigation button
and add it again when the back button is pressed. This way we no longer
have the problem the user going directly back to the main activity as
this is not what the user expects.

When pressing on DELETE in the ActionBar we now show an AlertDialog that
ask if we really want to delete the track. If the user accepts the track
is deleted and we go back to the track list.