Code reformat and indent fix for all files
Somehow some source files had a mixture of tabs and spaces used for
indention. I hopefully fixed this now. Also most of the layout files
were reformatted with code reformat. Maybe this can run automagically on
save.

Finally understood how ListFragements work. We now show a message if the
list is empty. We are also using the default progress bar that will pop
up while the CursorLoader is working.